+++ src/Doz.js
@@ -201,6 +201,7 @@ class Doz {
             });
         }
 
+        plugin.load(this);
         directive.callAppInit(this);
 
         if (this.cfg.mainComponent) {
@@ -234,8 +235,6 @@ class Doz {
 
         }
 
-        plugin.load(this);
-
         //Apply listeners
         if (this.cfg.listeners) {
             Object.keys(this.cfg.listeners).forEach(event => {
@@ -243,10 +242,6 @@ class Doz {
                 this.on(event, this.cfg.listeners[event])
             })
         }
-
-        //console.log('-----');
-
-        //directive.callAppInit(this);
 
         if (!this.cfg.mainComponent && this.cfg.autoDraw)
             this.draw();