+++ src/component/create-instance.js
@@ -16,7 +16,7 @@ function createInstance(cfg = {}) {
 
     if (!cfg.root) return;
 
-    if (!cfg.mountMainComponent) {
+    if (!(cfg.mountMainComponent || cfg.componentObject)) {
         if (cfg.template instanceof HTMLElement) {
             if (!cfg.template.parentNode)
                 cfg.root.appendChild(cfg.template);
@@ -249,14 +249,14 @@ function createInstance(cfg = {}) {
         }
     }
 
-    if (cfg.mountMainComponent) {
+    if (cfg.mountMainComponent || cfg.componentObject) {
         // Monto il componente principale
         let newElement = new cfg.component({
             //tag: 'bbb-bbb',//cmp.tag || cmpName,
             cmp: cfg.component,
             root: cfg.root,
             app: cfg.app,
-            props: {},
+            props: cfg.props || {},
             componentDirectives: {},
             parentCmp: null,//parent.cmp || cfg.parent
         });
@@ -265,7 +265,7 @@ function createInstance(cfg = {}) {
         newElement.app.emit('componentPropsInit', newElement);
 
         newElement._isRendered = true;
-        newElement._mainComponentByAppCreate = true;
+        newElement._mainComponentByAppCreate = !!cfg.mountMainComponent;
         newElement.render(true);
 
         if (cfg.innerHTML) {